requirements.txt
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
MouseInfo==0.1.3
oauthlib==3.2.2
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytweening==1.0.7
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.0
six==1.16.0
subprocess.run==0.0.8
tqdm==4.65.0
uritemplate==4.1.1
urllib3==1.26.15
